fuzzing: Put reader functions into separate file
