+++ src/tests/fuzzing/fuzz_pkcs15init.c
@@ -23,169 +23,12 @@
 #include "config.h"
 #endif
 
+#include "fuzzer_reader.h"
 #include "pkcs15init/pkcs15-lib.c"
 #include "scconf/scconf.h"
 #include "pkcs15init/pkcs15-init.h"
 #include "pkcs15init/profile.c"
 #include "pkcs15init/profile.h"
-#include "libopensc/internal.h"
-
-#include <stdlib.h>
-#include <string.h>
-
-
-
-/* Taken from fuzz_pkcs15_reader.c */
-struct driver_data {
-    const uint8_t *Data;
-    size_t Size;
-};
-
-static struct sc_reader_operations fuzz_ops = {0};
-static struct sc_reader_driver fuzz_drv = {
-    "Fuzzing reader",
-    "fuzz",
-    &fuzz_ops,
-    NULL
-};
-
-void fuzz_get_chunk(sc_reader_t *reader, const uint8_t **chunk, uint16_t *chunk_size)
-{
-    struct driver_data *data;
-    uint16_t c_size;
-    const uint8_t *c;
-
-    if (chunk)
-        *chunk = NULL;
-    if (chunk_size)
-        *chunk_size = 0;
-
-    if (!chunk || !chunk_size || !reader) {
-        sc_debug(reader->ctx, SC_LOG_DEBUG_VERBOSE_TOOL, "Invalid Arguments");
-        return;
-    }
-    data = reader->drv_data;
-    if (!data || !data->Data || data->Size < sizeof c_size) {
-        sc_debug(reader->ctx, SC_LOG_DEBUG_VERBOSE_TOOL, "Invalid Arguments");
-        return;
-    }
-    /* parse the length of the returned data on two bytes */
-    c_size = *((uint16_t *) data->Data);
-    /* consume two bytes from the fuzzing data */
-    data->Size -= sizeof c_size;
-    data->Data += sizeof c_size;
-
-    if (data->Size < c_size) {
-        c_size = data->Size;
-    }
-
-    /* consume the bytes from the fuzzing data */
-    c = data->Data;
-    data->Size -= c_size;
-    data->Data += c_size;
-
-    sc_debug_hex(reader->ctx, SC_LOG_DEBUG_VERBOSE_TOOL,
-        "Returning fuzzing chunk", c, c_size);
-
-    *chunk = c;
-    *chunk_size = c_size;
-}
-
-static int fuzz_reader_release(sc_reader_t *reader)
-{
-    if (reader) {
-        free(reader->drv_data);
-        reader->drv_data = NULL;
-    }
-
-    return SC_SUCCESS;
-}
-
-static int fuzz_reader_connect(sc_reader_t *reader)
-{
-    uint16_t chunk_size;
-    const uint8_t *chunk;
-
-    fuzz_get_chunk(reader, &chunk, &chunk_size);
-
-    if (chunk_size > SC_MAX_ATR_SIZE)
-        chunk_size = SC_MAX_ATR_SIZE;
-    else
-        reader->atr.len = chunk_size;
-
-    if (chunk_size > 0)
-        memcpy(reader->atr.value, chunk, chunk_size);
-
-    return SC_SUCCESS;
-}
-
-static int fuzz_reader_disconnect(sc_reader_t *reader)
-{
-    return SC_SUCCESS;
-}
-
-static int fuzz_reader_transmit(sc_reader_t *reader, sc_apdu_t *apdu)
-{
-    const uint8_t *chunk;
-    uint16_t chunk_size;
-
-    fuzz_get_chunk(reader, &chunk, &chunk_size);
-
-    if (chunk_size >= 2) {
-        /* set the SW1 and SW2 status bytes (the last two bytes of
-         * the response */
-        apdu->sw1 = (unsigned int)chunk[chunk_size - 2];
-        apdu->sw2 = (unsigned int)chunk[chunk_size - 1];
-        chunk_size -= 2;
-        /* set output length and copy the returned data if necessary */
-        if (chunk_size <= apdu->resplen)
-            apdu->resplen = chunk_size;
-
-        if (apdu->resplen != 0)
-            memcpy(apdu->resp, chunk, apdu->resplen);
-    } else {
-        apdu->sw1 = 0x6D;
-        apdu->sw2 = 0x00;
-        apdu->resplen = 0;
-    }
-
-    return SC_SUCCESS;
-}
-
-struct sc_reader_driver *sc_get_fuzz_driver(void)
-{
-    fuzz_ops.release = fuzz_reader_release;
-    fuzz_ops.connect = fuzz_reader_connect;
-    fuzz_ops.disconnect = fuzz_reader_disconnect;
-    fuzz_ops.transmit = fuzz_reader_transmit;
-    return &fuzz_drv;
-}
-
-void fuzz_add_reader(struct sc_context *ctx, const uint8_t *Data, size_t Size)
-{
-    sc_reader_t	*reader;
-    struct driver_data *data;
-    char name[64] = {0};
-
-    if (!(reader = calloc(1, sizeof(*reader)))
-            || !(data = (calloc(1, sizeof(*data))))) {
-        free(reader);
-        return;
-    }
-
-    data->Data = Data;
-    data->Size = Size;
-
-    reader->driver = &fuzz_drv;
-    reader->ops = &fuzz_ops;
-    reader->drv_data = data;
-    snprintf(name, sizeof name - 1, "%zu random byte%s reader (%p)",
-            Size, Size == 1 ? "" : "s", Data);
-    reader->name = strdup(name);
-
-    reader->ctx = ctx;
-    list_append(&ctx->readers, reader);
-}
 
 int fuzz_profile_load(struct sc_profile *profile, const uint8_t *data, size_t size)
 {
